requirements.txt
asgiref==3.5.2
black==22.10.0
click==8.1.3
colorama==0.4.5
Django==3.2.13
django-bootstrap5==22.1
mypy-extensions==0.4.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pytz==2022.4
python-dotenv==0.20.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0

certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
requests==2.28.1
urllib3==1.26.10